requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
matplotlib>=3.5.0
reportlab>=3.6.0
numpy>=1.21.0
